Dependecies, maven, new Files format added!
In this "DEAR COMMIT", first of all, I added maven, a dependency or lib manager...because I wanted to simplify my job, as an example, the getters and setters...make the code big.
Libs that I added are: Lombok, Lib4j and Gson.

Lombok: is for the getters and setter, makes the code lower.

Gson: Is for  manipulate ".Json" file, that I will use to store data and others stuff.

Lib4j: Is for creating Log for the program, making more Professional LOL.

and I will stop using ".txt" files, because for the data, I'll use ".Json", for   Program configuration will be ".properties", and for log, will be ".log"(I'm not sure yet)

I had optimized the code with minimum imports and stuff, and in the future...when I finish this... I'll compile the program to ".jar" file, and an Instaler