Program has been optimized!
In this commit, I had optimized the program, by removing useless classes and methods....an simplifing the code.

An I created the class "JsonFiles" that it saves and load the student and teacher data, and the class name will be changed  in the future.

In the future commit I will create a specific class for student data input and for the teacher too.

#BL4Z3:THE CREATOR!